third_party/blink/web_tests/editing/shadow/compare-positions-in-nested-shadow-expected.txt

CONSOLE WARNING: line 20: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS selection2.anchorNode is selection2.focusNode
PASS selection2.anchorOffset is 5
PASS selection2.focusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
